packages/rentl-llm/pyproject.toml
[project]
name = "rentl-llm"
version = "0.1.8"
description = "LLM runtime adapters for rentl"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "rentl-core",
    "rentl-schemas",
    "pydantic-ai>=1.47.0, <2",
]

[build-system]
requires = ["uv-build>=0.1.0"]
build-backend = "uv_build"

[tool.uv.build]
packages = ["src/rentl_llm"]
